# Reduce allocations/cpu from scraping