… go.sum
… vendor/github.com/valyala/fastjson/LICENSE
… vendor/github.com/valyala/fastjson/fastfloat/parse.go
… vendor/modules.txt